test: add stories

+++ bin/bench.sh
@@ -1,5 +1,9 @@
-luau --profile -O2 benchmarks/$1.luau
-python3 bin/perfgraph.py profile.out >profile.svg
-rm -rf bin/__pycache__
-rm profile.out
-open profile.svg
+if [ "$2" == "profile" ]; then
+	luau --profile -O2 benchmarks/$1.luau
+	python3 bin/perfgraph.py profile.out >profile.svg
+	rm -rf bin/__pycache__
+	rm profile.out
+	open profile.svg
+else
+	lute benchmarks/$1.luau
+fi
